Recombine partial classes to ease code review.
